Tremolo: Add SPICE simulation elements
The voltage sources and the model of RV2 are located outside / below the
sheet since they are not part of the "real" circuit. Set the voltage
sources and RV2 simulation-placeholder resister as being excluded from
the BOM and board. Set RV2 as excluded from the simulation since R7 and
R8 take its place.

Add another global label at the drain of Q2 so the two resistors that
model RV2 can be placed outside of the sheet.

Use 210mV @ 344Hz for the input per Linus's own testing.

The SPICE model for OPA1655 and PMV250 are expected to be inside the
Tremolo project directory, named "OPA1655.LIB" and
"PMV250EPEA_22_5_2013.lib" as they are from the manufacturers' websites.

Signed-off-by: Eric Kuzmenko <eric.n.kuzmenko@gmail.com>